[chore] #11 ViewHolder 분리
