+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/MyProfileViewHolder.kt
@@ -1,0 +1,32 @@
+package org.sopt.dosopttemplate.presentation.home
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import coil.load
+import org.sopt.dosopttemplate.databinding.ItemHomeMyProfileBinding
+import org.sopt.dosopttemplate.domain.model.Profile
+import org.sopt.dosopttemplate.presentation.type.ProfileInfoType
+
+class MyProfileViewHolder(
+    private val binding: ItemHomeMyProfileBinding
+) : RecyclerView.ViewHolder(binding.root) {
+    fun onBind(myProfile: Profile.MyProfile) {
+        binding.run {
+            ivMyProfile.load(myProfile.profileImage)
+            tvMyProfileName.text = myProfile.name
+            tvMyProfileDescription.text = myProfile.description
+            if (myProfile.description.isNullOrEmpty()) tvMyProfileDescription.visibility =
+                View.GONE
+
+            with(includeMyProfileInfo) {
+                layoutPortraitProfileInfo.setBackgroundResource(ProfileInfoType.DESCRIPTION.portraitBackgroundRes)
+                ProfileInfoType.DESCRIPTION.contextRes?.let {
+                    tvPortraitProfileInfoContext.setText(
+                        it
+                    )
+                }
+                ivPortraitProfileInfoIcon.setImageResource(ProfileInfoType.DESCRIPTION.iconRes)
+            }
+        }
+    }
+}
